server/main/requirements.txt
asgiref==3.8.1
boto3==1.35.25
botocore==1.35.25
certifi==2024.8.30
chardet==4.0.0
Django==5.1.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
fpdf==1.7.2
gunicorn==23.0.0
idna==2.10
jmespath==1.0.1
lxml==5.3.0
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.25.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
